Use the EOS ID directly
When adding documents, the EOS ID is used to mark the end of each of
them. The ID is acquired based on the string used to represent it, but
that differs from the default SentencePiece display string. However,
it's possible to get the EOS ID directly without reference to the
string, so this change just does that instead.